backend/pyproject.toml
[project]
name = "jeetu-code-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.83",
    "langchain-ollama>=0.3.10",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "scipy>=1.11.4",
    "numpy>=1.26.4",
    "sarvamai",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "sqlalchemy>=2.0.27",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.9",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "psycopg2-binary>=2.9.9",
    "langgraph>=0.6.11",
    "tavily-python>=0.7.21",
    "langchain-groq>=0.3.8",
    "langchain-postgres>=0.0.17",
    "psycopg[binary,pool]>=3.2.13",
    "pgvector>=0.3.6",
    "gunicorn>=23.0.0",
]
